# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Batchmates API",
        "description": "Multi-tenant alumni donation and campaign management system API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "List campaigns",
                "description": "Lists campaigns. Donors see only 'active' campaigns. Committee members see campaigns from their institution. Admins see all.",
                "operationId": "617d936d65b29bf3912371d823eceba3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "institution_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create a campaign request",
                "description": "Creates a new campaign. Status starts as 'pending_review' and requires committee approval to become 'active'.",
                "operationId": "c9e6f9a10cb181d5df074904ff19a753",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "goal_amount",
                                    "campaign_type",
                                    "institution_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Medical Assistance for Juan"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "goal_amount": {
                                        "type": "number",
                                        "example": 100000
                                    },
                                    "campaign_type": {
                                        "type": "string",
                                        "enum": [
                                            "general",
                                            "individual"
                                        ]
                                    },
                                    "institution_id": {
                                        "description": "Required. Institution this campaign belongs to.",
                                        "type": "integer"
                                    },
                                    "bank_account_id": {
                                        "description": "Bank account for this campaign",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "beneficiary_id": {
                                        "description": "For individual campaigns",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "urgent"
                                        ]
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign created successfully and pending review"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign details",
                "operationId": "c9abcff7d8ae0d70a68c8f917cef753d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign details"
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Update campaign",
                "description": "Creator can update pending campaigns. Committee/admin can update approved campaigns.",
                "operationId": "b021a1b170e89513ac38b1122d79d671",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "goal_amount": {
                                        "type": "number"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign updated"
                    },
                    "403": {
                        "description": "Cannot edit approved campaign"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Delete campaign",
                "description": "Creator can delete pending campaigns. Admin can delete any campaign.",
                "operationId": "22b8be238435eacfee5a7206cef05d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign deleted"
                    },
                    "403": {
                        "description": "Cannot delete approved campaign"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/campaignStats": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Campaign statistics",
                "description": "Get campaign statistics scoped to user's institution",
                "operationId": "6701d34b300c49442c02f7121bd3454a",
                "responses": {
                    "200": {
                        "description": "Campaign stats"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "description": "Use this for MOBILE testing (Token-based)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "CookieCsrfToken": {
                "type": "apiKey",
                "description": "Use this for WEB testing (Session-based). Value must be the XSRF-TOKEN cookie value.",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}